ISPN-14585 Convert RESP endpoint to use parser generator
* Refactor RespRequestHandler to not require Cache specific stuff
